UserService/requirements.txt
Flask
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
psycopg2-binary
python-dotenv
